+++ CHANGELOG.md
@@ -10,10 +10,16 @@ All notable changes to this project will be documented in this file. This change
 * core.matrix protocols for vectors
 * math functions can operate on vectors now (like [[sin]] etc.)
 * new functions for vectors `clamp`, `zero-count`, `nonzero-count`, `as-vec`
+* classification bindings to SMILE and XGBoost
 
 ### Changed
 
 * use `fmap` instead of `applyf` (now deprecated)
+* rbf function already implements Smile interface
+
+### Removed
+
+* rbf-obj converter
 
 ## [1.2.0]